requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements/base.txt requirements/base.in
#
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   google-genai
    #   httpx
    #   mcp
    #   sse-starlette
    #   starlette
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
authlib==1.5.2
    # via google-adk
blinker==1.9.0
    # via streamlit
cachetools==5.5.2
    # via
    #   google-auth
    #   streamlit
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   google-adk
    #   streamlit
    #   uvicorn
cryptography==44.0.2
    # via authlib
deprecated==1.2.18
    # via
    #   opentelemetry-api
    #   opentelemetry-semantic-conventions
docstring-parser==0.16
    # via google-cloud-aiplatform
exceptiongroup==1.2.2
    # via anyio
fastapi==0.115.12
    # via google-adk
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
google-adk==0.2.0
    # via -r requirements/base.in
google-api-core[grpc]==2.24.2
    # via
    #   google-api-python-client
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-resource-manager
    #   google-cloud-secret-manager
    #   google-cloud-speech
    #   google-cloud-storage
    #   google-cloud-trace
google-api-python-client==2.167.0
    # via google-adk
google-auth==2.39.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-resource-manager
    #   google-cloud-secret-manager
    #   google-cloud-speech
    #   google-cloud-storage
    #   google-cloud-trace
    #   google-genai
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-cloud-aiplatform==1.90.0
    # via google-adk
google-cloud-bigquery==3.31.0
    # via google-cloud-aiplatform
google-cloud-core==2.4.3
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
google-cloud-resource-manager==1.14.2
    # via google-cloud-aiplatform
google-cloud-secret-manager==2.23.3
    # via google-adk
google-cloud-speech==2.32.0
    # via google-adk
google-cloud-storage==2.19.0
    # via
    #   google-adk
    #   google-cloud-aiplatform
google-cloud-trace==1.16.1
    # via opentelemetry-exporter-gcp-trace
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-genai==1.11.0
    # via google-adk
google-resumable-media==2.7.2
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos[grpc]==1.70.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
graphviz==0.20.3
    # via google-adk
greenlet==3.2.1
    # via sqlalchemy
grpc-google-iam-v1==0.14.2
    # via
    #   google-cloud-resource-manager
    #   google-cloud-secret-manager
grpcio==1.72.0rc1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.71.0
    # via google-api-core
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.8
    # via httpx
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via
    #   google-genai
    #   mcp
httpx-sse==0.4.0
    # via mcp
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.6.1
    # via opentelemetry-api
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2025.4.1
    # via jsonschema
markupsafe==3.0.2
    # via jinja2
mcp==1.6.0
    # via google-adk
narwhals==1.36.0
    # via altair
numpy==2.2.5
    # via
    #   pandas
    #   pydeck
    #   shapely
    #   streamlit
opentelemetry-api==1.32.1
    # via
    #   google-adk
    #   opentelemetry-exporter-gcp-trace
    #   opentelemetry-resourcedetector-gcp
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-gcp-trace==1.9.0
    # via google-adk
opentelemetry-resourcedetector-gcp==1.9.0a0
    # via opentelemetry-exporter-gcp-trace
opentelemetry-sdk==1.32.1
    # via
    #   google-adk
    #   opentelemetry-exporter-gcp-trace
    #   opentelemetry-resourcedetector-gcp
opentelemetry-semantic-conventions==0.53b1
    # via opentelemetry-sdk
packaging==24.2
    # via
    #   altair
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   streamlit
pandas==2.2.3
    # via streamlit
pillow==11.2.1
    # via streamlit
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-resource-manager
    #   google-cloud-secret-manager
    #   google-cloud-speech
    #   google-cloud-trace
protobuf==5.29.4
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-resource-manager
    #   google-cloud-secret-manager
    #   google-cloud-speech
    #   google-cloud-trace
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
    #   streamlit
pyarrow==19.0.1
    # via streamlit
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.22
    # via cffi
pydantic==2.11.3
    # via
    #   -c /home/serafin/dev/personal-projects/spot2_assistant/requirements/constraints.txt
    #   -r requirements/base.in
    #   fastapi
    #   google-adk
    #   google-cloud-aiplatform
    #   google-genai
    #   mcp
    #   pydantic-settings
pydantic-core==2.33.1
    # via pydantic
pydantic-settings==2.9.1
    # via mcp
pydeck==0.9.1
    # via streamlit
pyparsing==3.2.3
    # via httplib2
python-dateutil==2.9.0.post0
    # via
    #   google-cloud-bigquery
    #   pandas
python-dotenv==1.1.0
    # via
    #   -r requirements/base.in
    #   google-adk
    #   pydantic-settings
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via google-adk
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   google-genai
    #   opentelemetry-resourcedetector-gcp
    #   streamlit
rpds-py==0.24.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
shapely==2.1.0
    # via google-cloud-aiplatform
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.40
    # via google-adk
sse-starlette==2.3.3
    # via mcp
starlette==0.46.2
    # via
    #   fastapi
    #   mcp
    #   sse-starlette
streamlit==1.44.1
    # via -r requirements/base.in
tenacity==9.1.2
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.4.2
    # via streamlit
typing-extensions==4.13.2
    # via
    #   altair
    #   anyio
    #   fastapi
    #   google-cloud-aiplatform
    #   google-genai
    #   opentelemetry-resourcedetector-gcp
    #   opentelemetry-sdk
    #   pydantic
    #   pydantic-core
    #   referencing
    #   sqlalchemy
    #   streamlit
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.0
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via pandas
tzlocal==5.3.1
    # via google-adk
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.4.0
    # via requests
uvicorn==0.34.2
    # via
    #   google-adk
    #   mcp
watchdog==6.0.0
    # via streamlit
websockets==15.0.1
    # via google-genai
wrapt==1.17.2
    # via deprecated
zipp==3.21.0
    # via importlib-metadata
